+++ routes/api.php
@@ -14,6 +14,8 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+Route::resource('/orders', 'OrderController');
+
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
